Add : private route added
