+++ src/Pages/LandingPage/LandingPage.jsx
@@ -9,38 +9,7 @@ import {
 } from "./LandingComponents/index";
 
 export const LandingPage = () => {
-  const { setWishlistProducts } = useWishlistContext();
-  const { setCartProducts } = useCartContext();
-  const encodedToken = localStorage.getItem("token");
-
-  useEffect(() => {
-    (async function () {
-      try {
-        const { data } = await axios.post("/api/auth/verify", {
-          encodedToken: encodedToken,
-        });
-
-        const cartResponse = await axios.get("/api/user/cart", {
-          headers: {
-            authorization: data.encodedToken,
-          },
-        });
-        if (cartResponse.status === 200) {
-          setCartProducts(cartResponse.data.cart);
-        }
-
-        const wishlistResponse = await axios.get("/api/user/wishlist", {
-          headers: {
-            authorization: data.encodedToken,
-          },
-        });
-        if (wishlistResponse.status === 200) {
-          setWishlistProducts(wishlistResponse.data.wishlist);
-        }
-      } catch (error) {}
-    })();
-  }, []);
-
+  
   return (
     <ShowcaseContainer>
       <TitleContainer />
